requirements.txt
annotated-types==0.7.0
ansicolors==1.1.8
anyio==4.6.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
distro==1.9.0
et-xmlfile==1.1.0
Flask==2.0.1
Flask-WTF==0.15.1
gitdb==4.0.7
GitPython==3.1.18
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.2
ijson==3.1.4
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
jiter==0.5.0
jsonpickle==2.0.0
Levenshtein==0.26.0
lizard==1.17.9
MarkupSafe==2.0.1
openai==1.50.0
openpyxl==3.0.7
pydantic==2.9.2
pydantic_core==2.23.4
PyDriller==1.15.5
pymongo==3.12.0
python-dateutil==2.8.2
python-Levenshtein==0.26.0
pytz==2021.1
RapidFuzz==3.10.0
redis==3.5.3
regex==2024.11.6
requests==2.26.0
setuptools==75.1.0
six==1.16.0
smmap==4.0.0
sniffio==1.3.1
tiktoken==0.8.0
tqdm==4.62.2
typing_extensions==4.12.2
unidiff==0.6.0
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
